wip: ported oversampling waveshaper

+++ Inc/audio/oversamplingWaveshaper.h
@@ -6,14 +6,14 @@
 
 
 typedef struct {
-    int16_t oldValue;
+    float oldValue;
     WaveShaperDataType waveshaper;
     SecondOrderIirFilterType oversamplingFilter;
 
 } OversamplingWaveshaperDataType;
 
 void initOversamplingWaveshaper(OversamplingWaveshaperDataType*data);
-void  applyOversamplingDistortion(uint16_t*data,OversamplingWaveshaperDataType*waveshaper);
-int16_t  OversamplingDistortionProcessSample(int16_t sample,OversamplingWaveshaperDataType* waveshaper);
+void  applyOversamplingDistortion(float*data,OversamplingWaveshaperDataType*waveshaper);
+float  OversamplingDistortionProcessSample(float sample,OversamplingWaveshaperDataType* waveshaper);
 void oversamplingWaveshaperReset(OversamplingWaveshaperDataType*data);
 #endif
